+++ Lab4/Battery.hpp
@@ -5,13 +5,13 @@
 
 class Battery : public Component {
 public:
-    Battery(std::string name, double voltage, Connection& a, Connection& b);
+    Battery(std::string name, float voltage, Connection& a, Connection& b);
     void simulate() const override;
     void current() override;
     void set_connection_values() const;
     float getcurrent() const override ;
 private:
-    double voltage{0.0};
+    float voltage{0.0};
     
 };
