+++ Lab4/Resistor.hpp
@@ -5,13 +5,13 @@
 
 class Resistor : public Component {
 public:
-    Resistor(std::string name, double resistance, Connection& a, Connection& b);
+    Resistor(std::string name, float resistance, Connection& a, Connection& b);
     void set_connection_values(float timestep);
     void simulate() const override;
     void current() override;
     float getcurrent() const override;
 private:
-    double const resistance{0.0};
+    float const resistance{0.0};
 };
 
 #endif
